Migrate from Integer to Long

+++ lib/src/main/java/org/seattleoba/data/model/stripe/BalanceTransaction.java
@@ -8,13 +8,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record BalanceTransaction(
         @JsonProperty("id") String id,
-        @JsonProperty("amount") Integer amount,
+        @JsonProperty("amount") Long amount,
         @JsonProperty("available_on") Long availableOn,
         @JsonProperty("created") Long created,
         @JsonProperty("currency") String currency,
         @JsonProperty("description") String description,
-        @JsonProperty("fee") Integer fee,
-        @JsonProperty("net") Integer net,
+        @JsonProperty("fee") Long fee,
+        @JsonProperty("net") Long net,
         @JsonProperty("source") String source,
         @JsonProperty("status") String status,
         @JsonProperty("type") String type) {
